still working on better parallel execution

+++ src/comorbidRPSetInt.cpp
@@ -95,7 +95,7 @@ struct ComorbidWorkerSetInt : public Worker {
 // [[Rcpp::export]]
 List icd9ComorbidShortRPSetInt(DataFrame icd9df, List icd9Mapping,
 		const std::string visitId = "visitId", const std::string icd9Field = "icd9",
-		size_t grainSize = 0) {
+		int grainSize = 0) {
 
 #ifdef ICD9_DEBUG
 	std::cout << "icd9ComorbidShortRPSetInt\n";
